+++ ext_tables.php
@@ -1,8 +1,8 @@
 <?php
 
-use JobRouter\Addon\Typo3Data\Controller\TableListController;
-use JobRouter\Addon\Typo3Data\Extension;
-use JobRouter\Addon\Typo3Data\Hooks\TableUpdateHook;
+use JobRouter\AddOn\Typo3Data\Controller\TableListController;
+use JobRouter\AddOn\Typo3Data\Extension;
+use JobRouter\AddOn\Typo3Data\Hooks\TableUpdateHook;
 use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;