+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Require JobRouter REST Client in version 3
-- Namespace from `Brotkrueml\JobRouterData` to `JobRouter\Addon\Typo3Data`
+- Namespace from `Brotkrueml\JobRouterData` to `JobRouter\AddOn\Typo3Data`
 
 ## [2.0.0] - 2023-05-31